requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.0
async-generator==1.10
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
bs4==0.0.2
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
datasets==4.5.0
dill==0.4.0
Django==5.2.10
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
filelock==3.20.3
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
outcome==1.3.0.post0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==3.0.0
playwright==1.57.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==23.0.0
pycparser==3.0
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.0.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.0
websocket-client==1.9.0
wsproto==1.3.2
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
